+++ Sources/SentryObjCCompat/SentryObjCOptions.swift
@@ -294,6 +294,20 @@ import Foundation
         set { wrapped.enablePersistingTracesWhenCrashing = newValue }
     }
 
+#if os(iOS) || os(macOS)
+    @objc public var configureProfiling: ((SentryObjCProfileOptions) -> Void)? {
+        didSet {
+            if let configureProfiling {
+                wrapped.configureProfiling = { profiling in
+                    configureProfiling(SentryObjCProfileOptions(profiling))
+                }
+            } else {
+                wrapped.configureProfiling = nil
+            }
+        }
+    }
+#endif // os(iOS) || os(macOS)
+
     @objc public var initialScope: ((SentryObjCScope) -> SentryObjCScope) = { return $0 } {
         didSet {
             let initialScope = initialScope